+++ x/profiles/types/events.go
@@ -27,8 +27,8 @@ const (
 	EventTypeBlockUser   = "block_user"
 	EventTypeUnblockUser = "unblock_user"
 
-	AttributeUserBlockBlocker = "blocker"
-	AttributeUserBlockBlocked = "blocked"
-	AttributeUserBlockReason  = "reason"
-	AttributeSubspace         = "subspace"
+	AttributeKeyUserBlockBlocker = "blocker"
+	AttributeKeyUserBlockBlocked = "blocked"
+	AttributeKeyUserBlockReason  = "reason"
+	AttributeKeySubspace         = "subspace"
 )